Manually forward `reactBridgeDidFinishTransaction`
As noted in https://github.com/brentvatne/react-native-modal/issues/11 the scrollview does not work in the overlay. This is due to the re-computation of the content size never happening. The scrollview computes the content size in the `reactBridgeDidFinishTransaction` method, which doesn't get called, because the `_overlayBaseView` is not registered as a root view with react.

We now forward the message manually to all subviews of the overlay view and thus allow the scrollview to work.